Potential fix for code scanning alert no. 12: Incorrect conversion between integer types

+++ common/services/notificationService.go
@@ -108,7 +108,7 @@ func HideNotificationByInvitationID(invitationID uint) {
 
 	if err != nil {
 		slog.Warn("Failed to hide notification for invitation",
-			slog.Int("invitation_id", int(invitationID)),
+			slog.Uint64("invitation_id", uint64(invitationID)),
 			slog.Any("error", err),
 		)
 	}
